+++ kubernetes/handler.go
@@ -14,6 +14,7 @@ func NewPluginWithService(service Service) *pluginbinding.Plugin {
 		pluginbinding.RegisterOperation(clusterListSpec(), service.ClusterList),
 		pluginbinding.RegisterOperation(clusterTestSpec(), service.ClusterTest),
 		pluginbinding.RegisterOperation(endpointDiscoverSpec(), service.EndpointDiscover),
+		pluginbinding.RegisterOperation(secretReadSpec(), service.SecretRead),
 		pluginbinding.RegisterOperation(namespaceListSpec(), service.NamespaceList),
 		pluginbinding.RegisterOperation(serviceListSpec(), service.ServiceList),
 		pluginbinding.RegisterOperation(serviceShowSpec(), service.ServiceShow),